refactor(docs): relocate badge utilities into docs tree
