+++ docs/_utils/__init__.py
@@ -1,0 +1,1 @@
+"""Helper utilities for documentation build."""
